Health UI + ReloadTime

+++ Assets/Scripts/Crate.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Crate : MonoBehaviour, IDamageable
 {
-    public void Damage(float amount)
+    public void Damage(int amount)
     {
         Debug.Log(amount + "U brake me, bro");
     }
